Refactor load_pbench_runs logic into various classes to improve extensibility
Created Abstract DiagnosticCheck class for providing template to add more checks
than the current ones we have for sosreports and controller_dir.
Created Concrete ControllerDirCheck and SosreportsCheck classes to refactor the
existing checks under this new system.
These enable increases extensibility and allow for easy addition of any more checks
and the option for the user to write their own checks.
Created PbenchRuns class to store various trackers for all the diagnostics used as
well as the run data itself. Added a run_diagnostic field to each run data stored
that stores the diagnostic information associated with each run (ie what was faulty
with the specific run if anything based on the checks used)