Add combine_data method to Collection class & attempt to merge 2 collection objects
Adds combine_data method to PbenchCombinedDataCollection class that takes
in another such object and adds all the data from the collection passed in
to the collection this method called from.
Creates method to test whether the calling of this function works as desired.
Goal: to have separate processes generate their own PbenchCombinedDataCollection
objects and then to merge all the data together.