Implement multiprocessing capability
Uses pathos because dill can better serialize objects. Creates
a ProcessPool and stores it in an attribute of collection object.
Makes Collection object take in no of cpus to use in initialization,
so pool can be initialized. Adds methods to add one month asynchornously,
or a list of months. Add methods to merge dictionaries in a special way,
and combine another Collection object's data into its own. This because
self is returned from the worker processes.